implement explosion cleanup system

+++ src/features/util.rs
@@ -1,5 +1,5 @@
 use crate::constants::FIXED_UPDATE_FREQUENCY;
 
-pub fn seconds_to_freq(seconds: i32) -> i32 {
-    seconds * FIXED_UPDATE_FREQUENCY as i32
+pub fn seconds_to_freq(seconds: f32) -> i32 {
+    (seconds * FIXED_UPDATE_FREQUENCY as f32) as i32
 }
